fix: remove annoying and useless warnings
When running revel locally we get tons of warnings about structs that are not embedding from revel.Controller. But we do this on purpose and the warnings are of no value to us. Now that we have our own forked version of revel/cmd we might as well remove the warnings.